Refactor user conflict handling in roster management and improve bulk insert logic
