+++ app/models/rosters/user_already_in_bundle_error.rb
@@ -1,0 +1,10 @@
+module Rosters
+  class UserAlreadyInBundleError < StandardError
+    attr_reader :conflicting_group
+
+    def initialize(conflicting_group)
+      @conflicting_group = conflicting_group
+      super
+    end
+  end
+end
